Fetch card details (masked)
