+++ packages/wallet/backend/src/gatehub/client.ts
@@ -36,10 +36,10 @@ import {
   IGetTransactionsResponse
 } from '@wallet/shared/src'
 import { BadRequest } from '@shared/backend'
+import type { ICardResponse } from '@wallet/shared'
 import {
   ICardDetailsResponse,
   ILinksResponse,
-  ICardResponse,
   ICreateCustomerRequest,
   ICreateCustomerResponse,
   ICardProductResponse,
