+++ packages/wallet/frontend/src/components/userCards/UserCard.tsx
@@ -1,11 +1,11 @@
-import { useState, type ComponentProps } from 'react'
+import { useMemo, useState, type ComponentProps } from 'react'
 import { CopyButton } from '@/ui/CopyButton'
 import { Chip, GateHubLogo, MasterCardLogo } from '../icons/UserCardIcons'
 import { cn } from '@/utils/helpers'
-import type { IUserCard } from '@/lib/api/card'
 import { useCardContext, UserCardContext } from './UserCardContext'
 import { UserCardActions } from './UserCardActions'
 import { UserCardSettings } from './UserCardSettings'
+import type { ICardResponse } from '@wallet/shared'
 
 export type UserCardContainerProps = ComponentProps<'div'>
 
@@ -28,13 +28,15 @@ const UserCardContainer = ({
 }
 
 interface UserCardFrontProps extends ComponentProps<'div'> {
-  card: IUserCard
+  nameOnCard: ICardResponse['nameOnCard']
+  isBlocked: boolean
 }
 
 // Even if the UserCard lives inside the context we explicitly pass the card
 // details as a prop, since we have to use this component in dialogs as well.
 export const UserCardFront = ({
-  card,
+  nameOnCard,
+  isBlocked,
   className,
   ...props
 }: UserCardFrontProps) => {
@@ -43,7 +45,7 @@ export const UserCardFront = ({
       <div
         className={cn(
           'flex flex-col h-full',
-          card.isFrozen ? 'select-none pointer-events-none blur' : ''
+          isBlocked ? 'select-none pointer-events-none blur' : ''
         )}
       >
         <div className="flex justify-between text-sm items-center">
@@ -54,11 +56,11 @@ export const UserCardFront = ({
           <Chip />
         </div>
         <div className="flex mt-auto justify-between items-center">
-          <span className="uppercase opacity-50">{card.name}</span>
+          <span className="uppercase opacity-50">{nameOnCard}</span>
           <MasterCardLogo />
         </div>
       </div>
-      {card.isFrozen ? (
+      {isBlocked ? (
         <div className="absolute inset-0 z-10 bg-[url('/frozen.webp')] bg-cover bg-center opacity-50" />
       ) : null}
     </UserCardContainer>
@@ -111,19 +113,27 @@ const UserCardBack = () => {
 }
 
 interface UserCardProps {
-  card: IUserCard
+  card: ICardResponse
 }
 export const UserCard = ({ card }: UserCardProps) => {
   const [showDetails, setShowDetails] = useState(false)
+
+  // Q: When the user freezes the card is it going to have the status `Blocked`
+  // or `TemporaryBlocked`
+  //
+  // Q; When the user receives the card, what status will it have?
+  const isBlocked = useMemo(() => card.status === 'Blocked', [card])
 
   return (
     <UserCardContext.Provider value={{ card, showDetails, setShowDetails }}>
       <div className="grid grid-cols-1 md:grid-cols-[20rem_1fr] max-w-3xl gap-x-24">
         <div className="space-y-6 max-w-80 mx-auto">
-          {card.isFrozen ? <UserCardFront card={card} /> : null}
-          {!card.isFrozen && showDetails ? <UserCardBack /> : null}
-          {!card.isFrozen && !showDetails ? (
-            <UserCardFront card={card} />
+          {isBlocked ? (
+            <UserCardFront nameOnCard={card.nameOnCard} isBlocked={isBlocked} />
+          ) : null}
+          {!isBlocked && showDetails ? <UserCardBack /> : null}
+          {!isBlocked && !showDetails ? (
+            <UserCardFront nameOnCard={card.nameOnCard} isBlocked={isBlocked} />
           ) : null}
           <UserCardActions />
         </div>
