+++ packages/wallet/shared/src/types/card.ts
@@ -62,3 +62,19 @@ export interface IGetTransactionsResponse {
   data: ITransaction[]
   pagination: IPagination
 }
+
+export interface ICardResponse {
+  sourceId: string
+  nameOnCard: string
+  productCode: string
+  id: string
+  accountId: string
+  accountSourceId: string
+  maskedPan: string
+  status: string
+  statusReasonCode: string | null
+  lockLevel: string | null
+  expiryDate: string
+  customerId: string
+  customerSourceId: string
+}
